% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{an.max}
\alias{an.max}
\title{Calculate annual maximum}
\usage{
an.max(data)
}
\arguments{
\item{data}{A numeric vector of monthly data.}
}
\value{
A numeric vector of annual maximums.
}
\description{
This function calculates the annual maximum from monthly data.
}
