% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{an.mean}
\alias{an.mean}
\title{Calculate annual mean}
\usage{
an.mean(data)
}
\arguments{
\item{data}{A numeric vector of monthly data.}
}
\value{
A numeric vector of annual means.
}
\description{
This function calculates the annual mean from monthly data.
}
