% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{an.sum}
\alias{an.sum}
\title{Calculate annual sum}
\usage{
an.sum(data)
}
\arguments{
\item{data}{A numeric vector of monthly data.}
}
\value{
A numeric vector of annual sums.
}
\description{
This function calculates the annual sum from monthly data.
}
