% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{as_raster}
\alias{as_raster}
\title{Convert vector to raster}
\usage{
as_raster(x, nrows = 360, ncols = 720)
}
\arguments{
\item{x}{A numeric vector.}

\item{nrows}{Number of rows in the raster (default is 360).}

\item{ncols}{Number of columns in the raster (default is 720).}
}
\value{
A raster object.
}
\description{
This function converts a vector to a raster format.
}
