% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{as_stack}
\alias{as_stack}
\title{Convert vector to raster stack}
\usage{
as_stack(x, nrows = 360, ncols = 720)
}
\arguments{
\item{x}{A numeric vector.}

\item{nrows}{Number of rows in the raster (default is 360).}

\item{ncols}{Number of columns in the raster (default is 720).}
}
\value{
A raster stack object.
}
\description{
This function converts a vector to a raster stack format.
}
