% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{deseason_add}
\alias{deseason_add}
\title{Deseasonalize data by subtracting monthly mean}
\usage{
deseason_add(data)
}
\arguments{
\item{data}{A numeric vector of monthly data.}
}
\value{
A deseasonalized numeric vector.
}
\description{
This function removes seasonal effects by subtracting the monthly mean.
}
