% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{deseason_mul}
\alias{deseason_mul}
\title{Deseasonalize data by dividing by monthly mean}
\usage{
deseason_mul(data)
}
\arguments{
\item{data}{A numeric vector of monthly data.}
}
\value{
A deseasonalized numeric vector.
}
\description{
This function removes seasonal effects by dividing by the monthly mean.
}
