% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{remove_except}
\alias{remove_except}
\title{Remove all objects except specified ones}
\usage{
remove_except(keep_vars)
}
\arguments{
\item{keep_vars}{A character vector of variable names to keep.}
}
\description{
This function removes all objects from the global environment except those specified.
}
