% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from numeric vector}
\usage{
remove_outliers(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A numeric vector with outliers replaced by NA.
}
\description{
This function replaces outliers in a numeric vector with NA.
}
