% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{remove_outliers_df}
\alias{remove_outliers_df}
\title{Remove outliers from data frame column}
\usage{
remove_outliers_df(data, column)
}
\arguments{
\item{data}{A data frame.}

\item{column}{The column name as a string.}
}
\value{
A data frame with outliers removed.
}
\description{
This function removes outliers from a specified column in a data frame.
}
