% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{zscore}
\alias{zscore}
\title{Normalize data using Z-score}
\usage{
zscore(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A numeric vector of Z-scores.
}
\description{
This function normalizes the data to a Z-score.
}
